+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Change History
 Version 0.8.5 (pending)
 - MSO role permissions cleanup
+- Fixed pronouns in staff import
 
 Version 0.8.4 (10/26/2019)
 - Added attendance report